+++ nemo/collections/llm/gpt/model/mistral_7b.py
@@ -10,7 +10,7 @@ from typing_extensions import Annotated
 from nemo.collections.llm.gpt.model.base import GPTConfig, GPTModel
 from nemo.collections.llm.utils import Config
 from nemo.lightning import io, teardown
-from nemo.lightning.pytorch.opt import OptimizerModule
+from nemo.lightning.pytorch.optim import OptimizerModule
 
 if TYPE_CHECKING:
     from transformers import MistralConfig, MistralForCausalLM
