+++ nemo/collections/llm/gpt/model/mixtral.py
@@ -7,7 +7,7 @@ import torch.nn.functional as F
 
 from nemo.collections.llm.gpt.model.base import GPTConfig, GPTModel
 from nemo.lightning import io, teardown
-from nemo.lightning.pytorch.opt import OptimizerModule
+from nemo.lightning.pytorch.optim import OptimizerModule
 
 if TYPE_CHECKING:
     from transformers import MistralConfig, MistralForCausalLM
